fixes and tests

+++ NSL.StaticWebStorage/Services/BasicTokensService.cs
@@ -9,38 +9,88 @@ namespace NSL.StaticWebStorage.Services
     {
         public abstract string StoragePath { get; }
 
-        public record StorageToken(StorageTokenModel? token, DateTime loadTime);
-
-        private ConcurrentDictionary<string, StorageToken> storage = new ConcurrentDictionary<string, StorageToken>();
+        private ConcurrentDictionary<string, TempTokenData> storage = new ConcurrentDictionary<string, TempTokenData>();
 
         protected StaticStorageConfigurationModel Configuration => configurationOptions.CurrentValue;
 
-        public StorageTokenModel? TryGetToken(string token)
+        public StorageTokenModel? TryGetToken(string? token)
         {
+            if (token == null)
+                return default;
+
             token = token.ToLower().Trim();
 
             var info = storage.GetOrAdd(token, token =>
             {
-                var epath = Path.Combine(StoragePath, token);
+                var epath = StoragePath;
+
+                epath = Path.Combine(epath, token);
 
                 if (!File.Exists(epath))
-                    return new StorageToken(null, DateTime.UtcNow);
+                {
+                    if (Configuration.MasterTokens.TryGetValue(token, out var code))
+                    {
+                        return new TempTokenData()
+                        {
+                            Token = new StorageTokenModel
+                            {
+                                Code = code,
+                                CanDownload = true,
+                                CanUpload = true,
+                                CanShareAccess = true,
+                            }
+                        };
+                    }
 
-                return new StorageToken(JsonSerializer.Deserialize<StorageTokenModel>(File.ReadAllText(epath), JsonSerializerOptions.Web), DateTime.UtcNow);
+                    return new();
+                }
+
+                return new()
+                {
+                    Token = JsonSerializer.Deserialize<StorageTokenModel>(File.ReadAllText(epath), JsonSerializerOptions.Web)
+                };
             });
 
-            return info.token;
+            info.AccessTime = DateTime.UtcNow;
+
+            return info.Token;
         }
 
 
-        protected bool CreateToken(string token, StorageTokenModel data)
+        public bool TryRemoveToken(string token)
         {
-            if(!Directory.Exists(StoragePath))
-                Directory.CreateDirectory(StoragePath);
+            token = token.ToLower().Trim();
 
-            var epath = Path.Combine(StoragePath, token);
+            var epath = StoragePath;
+
+            epath = Path.Combine(epath, token);
+
+            if (File.Exists(epath))
+            {
+                storage.TryRemove(token, out _);
+                File.Delete(epath);
+            }
+
+            return true;
+        }
+
+
+        public bool CreateToken(string token, StorageTokenModel data)
+        {
+            var epath = StoragePath;
+
+            if (!Directory.Exists(epath))
+                Directory.CreateDirectory(epath);
+
+            epath = Path.Combine(epath, token);
+
+            if (File.Exists(epath))
+                return false;
 
             File.WriteAllText(epath, JsonSerializer.Serialize(data, JsonSerializerOptions.Web));
+
+            storage.TryRemove(token, out _);
+
 
             return true;
         }
